+++ tests/BusinessLogic/WebhookAPI/MockComponents/MockShopErrorOrderService.php
@@ -2,6 +2,7 @@
 
 namespace SeQura\Core\Tests\BusinessLogic\WebhookAPI\MockComponents;
 
+use SeQura\Core\BusinessLogic\Domain\Webhook\Models\Webhook;
 use SeQura\Core\BusinessLogic\Webhook\Services\ShopOrderService;
 use SeQura\Core\Infrastructure\Http\Exceptions\HttpRequestException;
 
@@ -10,7 +11,7 @@ class MockShopErrorOrderService implements ShopOrderService
     /**
      * @throws HttpRequestException
      */
-    public function updateStatus(string $orderId, string $status)
+    public function updateStatus(Webhook $webhook, string $status)
     {
         throw new HttpRequestException('Error updating shop order status.');
     }