+++ tests/BusinessLogic/WebhookAPI/MockComponents/MockShopOrderService.php
@@ -2,6 +2,7 @@
 
 namespace SeQura\Core\Tests\BusinessLogic\WebhookAPI\MockComponents;
 
+use SeQura\Core\BusinessLogic\Domain\Webhook\Models\Webhook;
 use SeQura\Core\BusinessLogic\Webhook\Services\ShopOrderService;
 
 class MockShopOrderService implements ShopOrderService
@@ -9,7 +10,7 @@ class MockShopOrderService implements ShopOrderService
     /**
      * @inheritDoc
      */
-    public function updateStatus(string $orderId, string $status)
+    public function updateStatus(Webhook $webhook, string $status)
     {
     }
 }